+++ packages/kiarina-agi-image/src/kiarina/agi/image_generation_provider_impl/google/_models/google_image_generation_provider.py
@@ -30,10 +30,6 @@ logger = logging.getLogger(__name__)
 
 
 class GoogleImageGenerationProvider(BaseImageGenerationProvider):
-    """
-    Google Gemini Image Provider Implementation
-    """
-
     def __init__(self, settings: GoogleImageGenerationProviderSettings) -> None:
         super().__init__()
 
@@ -44,24 +40,13 @@ class GoogleImageGenerationProvider(BaseImageGenerationProvider):
         return f"{self.__class__.__name__}({self.settings.model_name})"
 
     @property
-    def google_auth_settings(self) -> kiarina.lib.google.GoogleSettings:
-        return kiarina.lib.google.settings_manager.get_settings(
-            self.settings.google_auth_settings_key
-        )
-
-    @property
-    def credentials(self) -> kiarina.lib.google.Credentials:
-        return kiarina.lib.google.get_credentials(
-            settings=self.google_auth_settings,
-            scopes=[
-                "https://www.googleapis.com/auth/cloud-platform",
-            ],
-        )
-
-    @property
     def client(self) -> genai.Client:
         if self._client is None:
-            self._client = genai.Client(credentials=self.credentials)
+            self._client = genai.Client(
+                **kiarina.lib.google.get_genai_options(
+                    self.settings.google_auth_settings_key,
+                )
+            )
 
         return self._client