+++ packages/kiarina-agi-image/tests/image_embedding_provider_impl/gemini/_helpers/test_create_gemini_image_embedding_provider.py
@@ -4,8 +4,5 @@ from kiarina.agi.image_embedding_provider_impl.gemini import (
 
 
 def test_create_gemini_image_embedding_provider() -> None:
-    _ = create_gemini_image_embedding_provider(
-        backend_type="gemini_api",
-        google_auth_settings_key="api_key",
-    )
+    _ = create_gemini_image_embedding_provider()
     assert True