+++ packages/kiarina-agi-text/src/kiarina/agi/text_embedding_provider_impl/google/_types/backend_type.py
@@ -1,7 +1,0 @@
-from typing import Literal, TypeAlias
-
-BackendType: TypeAlias = Literal[
-    "gemini_api",
-    "vertex_ai",
-    "vertex_ai_api_key",
-]