+++ packages/kiarina-lib-google/src/kiarina/lib/google/__init__.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from ._helpers.get_credentials import get_credentials
+    from ._helpers.get_genai_options import get_genai_options
     from ._helpers.get_self_signed_jwt import get_self_signed_jwt
     from ._settings import GoogleSettings, settings_manager
     from ._types.credentials import Credentials
@@ -20,6 +21,7 @@ __version__ = version("kiarina-lib-google")
 __all__ = [
     # ._helpers
     "get_credentials",
+    "get_genai_options",
     "get_self_signed_jwt",
     # ._settings
     "GoogleSettings",
@@ -45,6 +47,7 @@ def __getattr__(name: str) -> object:
     module_map = {
         # ._helpers
         "get_credentials": "._helpers.get_credentials",
+        "get_genai_options": "._helpers.get_genai_options",
         "get_self_signed_jwt": "._helpers.get_self_signed_jwt",
         # ._settings
         "GoogleSettings": "._settings",